aula 3

+++ src/routes/memories.ts
@@ -2,9 +2,17 @@
 import { FastifyInstance } from 'fastify';
 import { prisma } from './lib/prisma';
 import { z } from 'zod';
+
 export async function memoriesRoutes(app: FastifyInstance){
-  app.get('/memories', async ()=>{
+  app.addHook('preHandler', async (request)=>{
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request)=>{
     const memories = await prisma.memory.findMany({
+      where:{
+        userId: request.user.sub,
+      },
       orderBy:{
         createdAt: 'asc',
       }
@@ -19,7 +27,7 @@ export async function memoriesRoutes(app: FastifyInstance){
     })
   })
 
-  app.get('/memories/:id', async (request)=>{
+  app.get('/memories/:id', async (request, reply)=>{
    const paramsSchema = z.object({
     id: z.string().uuid(),
    })
@@ -31,6 +39,10 @@ export async function memoriesRoutes(app: FastifyInstance){
       id,
     }
    })
+
+   if(!memory.isPublic && memory.userId !== request.user.sub){
+    return reply.status(401).send()
+   }
 
    return memory
   })
@@ -49,14 +61,14 @@ export async function memoriesRoutes(app: FastifyInstance){
         content,
         coverUrl,
         isPublic,
-        userId: 'cef5280c-3a40-485c-9876-5487a5c4c455'
+        userId: request.user.sub
       }
      })
 
      return memory
   })
 
-  app.put('/memories/:id', async (request)=>{
+  app.put('/memories/:id', async (request,reply)=>{
     const paramsSchema = z.object({
       id: z.string().uuid(),
      })
@@ -70,6 +82,15 @@ export async function memoriesRoutes(app: FastifyInstance){
      })
 
      const {content, coverUrl, isPublic} = bodySchema.parse(request.body)
+
+     let memory = await prisma.memory.findFirstOrThrow({
+      where:{
+        id,
+      }
+     })
+     if(memory.userId !== request.user.sub){
+      return reply.status(401).send()
+     }
 
      const memomy = await prisma.memory.update({
       where:{
@@ -85,12 +106,21 @@ export async function memoriesRoutes(app: FastifyInstance){
      return memomy
   })
 
-  app.delete('/memories/:id', async (request)=>{
+  app.delete('/memories/:id', async (request,reply)=>{
     const paramsSchema = z.object({
       id: z.string().uuid(),
      })
   
      const {id} = paramsSchema.parse(request.params)
+
+     const memory = await prisma.memory.findFirstOrThrow({
+      where:{
+        id,
+      }
+     })
+     if(memory.userId !== request.user.sub){
+      return reply.status(401).send()
+     }
   
      await prisma.memory.delete({
       where:{
